Updates: - Fix docblocks - Refactor Slim\Container injection in class __construct to ContainerAware trait - Concat all middlewares into single class and simplify its usage
